requirements.txt
fastapi
uvicorn
python-multipart
notion-client
pydantic-settings
requests
python-docx
mammoth
beautifulsoup4
google-genai
sib-api-v3-sdk